[REFACTOR] 성능 개선 : 컴포넌트 분리 및 memo
