+++ src/components/layout/side-menu.jsx
@@ -36,4 +36,4 @@ const SideMenu = ({ largeCate, smallCates }) => {
   );
 };
 
-export default SideMenu;
+export default React.memo(SideMenu);
